Validated adding location to the advertisement

+++ search-service/src/main/java/com/uns/ftn/searchservice/domain/Vehicle.java
@@ -24,7 +24,7 @@ public class Vehicle {
     private Long id;
 
     @Column(name = "kilometersTraveled")
-    private double kilometersTraveled;
+    private int kilometersTraveled;
 
     @Column(name = "childSeatNumber")
     private int childSeatNumber;
